requirements.txt
fastapi
uvicorn
transformers
torch
streamlit
scikit-learn
numpy
requests
python-dotenv